api/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
sqlalchemy==2.0.32
psycopg[binary]==3.2.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==1.0.1
neo4j==5.22.0
httpx==0.27.0
email-validator==2.1.0.post1
requests==2.32.3
